Fix move to non-existing directories in -A
Required folders are now automatically created before trying to move a file from
$HOME to the stage when adding new files. Additionally, the add() function does
no longer halt execution when the file has already been moved and linked, which
allows -Ac to automatically commit a previously added (with -A only) file. This
behavior is more consistent with -G, where the same workflow is also possible.